changing the type of n in the print_num function

+++ main.h
@@ -9,7 +9,7 @@
 int check_type(char *ptr, va_list arg);
 int print_str(char *s);
 int _putchar(int c);
-int print_num(int n, int base);
+int print_num(long n, int base);
 int _printf(char *format, ...);
 
 #endif
